+++ src/main/java/nextstep/courses/domain/session/SessionPrice.java
@@ -4,20 +4,20 @@ public class SessionPrice {
 
     private static final int ZERO_PRICE = 0;
 
-    private int price;
+    private Long price;
 
-    public SessionPrice(int price) {
+    public SessionPrice(Long price) {
         validate(price);
         this.price = price;
     }
 
-    private void validate(int price) {
+    private void validate(Long price) {
         if(price < ZERO_PRICE) {
             throw new IllegalArgumentException("금액은 음수일 수 없습니다");
         }
     }
 
-    public boolean validatePrice(int userPay) {
+    public boolean validatePrice(Long userPay) {
         if(this.price != userPay) {
             throw new IllegalArgumentException("정확한 수강료를 납부해주세요");
         }
